Update main.py: "Set Union"
When working with set or frozenset container, one of the most common operations we can perform is a merge. To do this, we can return the union of two sets using the .union() method or | operator. Doing so will return a new set or frozenset containing all elements from both sets without duplicates. 

Adds new thing, user hashtags and song tags, added by an artist. Will be expanded once songlist is bigger.